requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.16,<3.17
django-unfold>=0.22,<0.23
django-celery-beat>=2.7,<2.8
markdown>=3.5,<3.6
Pillow>=11.0,<11.1 